Move Actions to admin's components
